# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowWeave Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/node-directories": {
      "get": {
        "summary": "Directories",
        "operationId": "directories_api_v1_node_directories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Directories Api V1 Node Directories Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Directory",
        "operationId": "create_directory_api_v1_node_directories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Directory Api V1 Node Directories Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/node-assets": {
      "get": {
        "summary": "Assets",
        "operationId": "assets_api_v1_node_assets_get",
        "parameters": [
          {
            "name": "directory_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Directory Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Assets Api V1 Node Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Asset",
        "operationId": "create_asset_api_v1_node_assets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeAssetWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Asset Api V1 Node Assets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Assets",
        "operationId": "delete_assets_api_v1_node_assets_delete",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeAssetBulkDeleteWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Assets Api V1 Node Assets Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/node-assets/{asset_id}": {
      "get": {
        "summary": "Asset",
        "operationId": "asset_api_v1_node_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Asset Api V1 Node Assets  Asset Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Asset",
        "operationId": "update_asset_api_v1_node_assets__asset_id__put",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeAssetWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Asset Api V1 Node Assets  Asset Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Asset",
        "operationId": "delete_asset_api_v1_node_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capability-imports/validate": {
      "post": {
        "summary": "Validate Capability",
        "operationId": "validate_capability_api_v1_capability_imports_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityValidateWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Capability Api V1 Capability Imports Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capability-imports": {
      "post": {
        "summary": "Commit Capability",
        "operationId": "commit_capability_api_v1_capability_imports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityCommitWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Commit Capability Api V1 Capability Imports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Capabilities Api V1 Capabilities Get"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Capability Batch",
        "operationId": "delete_capability_batch_api_v1_capabilities_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityBulkDeleteWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Delete Capability Batch Api V1 Capabilities Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities/{capability_id}/source": {
      "get": {
        "summary": "Capability Source",
        "operationId": "capability_source_api_v1_capabilities__capability_id__source_get",
        "parameters": [
          {
            "name": "capability_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Capability Source Api V1 Capabilities  Capability Id  Source Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Capability Source",
        "operationId": "update_capability_source_api_v1_capabilities__capability_id__source_put",
        "parameters": [
          {
            "name": "capability_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilitySkillRevisionWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Capability Source Api V1 Capabilities  Capability Id  Source Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities/{capability_id}": {
      "delete": {
        "summary": "Delete Capability",
        "operationId": "delete_capability_api_v1_capabilities__capability_id__delete",
        "parameters": [
          {
            "name": "capability_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilityBulkDeleteWrite": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "CapabilityBulkDeleteWrite"
      },
      "CapabilityCommitWrite": {
        "properties": {
          "import_token": {
            "type": "string",
            "title": "Import Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "import_token"
        ],
        "title": "CapabilityCommitWrite"
      },
...
      "CapabilitySkillRevisionWrite": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 1048576,
            "minLength": 1,
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content"
        ],
        "title": "CapabilitySkillRevisionWrite"
      },
      "CapabilityValidateWrite": {
        "properties": {
          "capability_type": {
            "type": "string",
            "enum": [
              "SKILL",
              "MCP",
              "HOOK"
            ],
            "title": "Capability Type"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_base64": {
            "type": "string",
            "title": "Content Base64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capability_type",
          "filename",
          "content_base64"
        ],
        "title": "CapabilityValidateWrite"
      },
      "CapabilityWrite": {
        "properties": {
          "capability_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 48,
                "minLength": 38
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Id"
          },
          "capability_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "SKILL",
                  "MCP",
                  "HOOK"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Type"
          },
          "capability_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Key"
          },
          "normalized_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CapabilityWrite"
      },
...
      "DirectoryWrite": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DirectoryWrite"
      },
...
      "ExecutorWrite": {
        "properties": {
          "model_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider Id"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "startup_prompt": {
            "type": "string",
            "title": "Startup Prompt",
            "default": ""
          },
          "context_prompt": {
            "type": "string",
            "title": "Context Prompt",
            "default": ""
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 900
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExecutorWrite"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IOFieldWrite": {
        "properties": {
          "field_key": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_]{0,99}$",
            "title": "Field Key"
          },
          "display_name": {
            "type": "string",
            "maxLength": 160,
            "title": "Display Name",
            "default": ""
          },
          "data_type": {
            "type": "string",
            "const": "URL",
            "title": "Data Type",
            "default": "URL"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "template_url": {
            "type": "string",
            "title": "Template Url",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field_key"
        ],
        "title": "IOFieldWrite"
      },
...
      "NodeAssetBulkDeleteWrite": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "NodeAssetBulkDeleteWrite"
      },
      "NodeAssetWrite": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "directory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Id"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "icon_kind": {
            "type": "string",
            "title": "Icon Kind",
            "default": "LUCIDE"
          },
          "icon_value": {
            "type": "string",
            "title": "Icon Value",
            "default": "bot"
          },
          "environment_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment Version Id"
          },
          "row_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Version"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/IOFieldWrite"
            },
            "type": "array",
            "title": "Inputs"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/IOFieldWrite"
            },
            "type": "array",
            "title": "Outputs"
          },
          "executor": {
            "$ref": "#/components/schemas/ExecutorWrite"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityWrite"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NodeAssetWrite"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}